# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BHUVISION \u2014 Agentic Earth Intelligence",
    "description": "An Interactive Vision-Language Assistant for Multimodal Remote Sensing Image Analysis through Text Queries (SIH26167, ISRO). Powered by 9 autonomous specialist agents.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/futuristic/sam-geo/segment": {
      "get": {
        "tags": [
          "Futuristic Earth Intelligence"
        ],
        "summary": "Segment Anything Geospatial",
        "description": "Simulates Segment Anything for Geospatial (SAM-Geo / SAM 2) zero-shot foundation segmentation.\nExtracts individual building footprints, road vectors, and water bodies from natural language prompts.",
        "operationId": "segment_anything_geospatial_api_futuristic_sam_geo_segment_get",
        "parameters": [
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Text or point prompt for SAM-Geo foundation model",
              "default": "all structures and water bodies",
              "title": "Prompt"
            },
            "description": "Text or point prompt for SAM-Geo foundation model"
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 12.9716,
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 77.5946,
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Segment Anything Geospatial Api Futuristic Sam Geo Segment Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/futuristic/insar/subsidence": {
      "get": {
        "tags": [
          "Futuristic Earth Intelligence"
        ],
        "summary": "Get Insar Subsidence Profile",
        "description": "Returns Synthetic Aperture Radar Interferometry (InSAR) millimeter-scale ground displacement time-series.\nUnwraps phase differentials between Sentinel-1 / RISAT ascending and descending passes.",
        "operationId": "get_insar_subsidence_profile_api_futuristic_insar_subsidence_get",
        "parameters": [
          {
            "name": "location_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Target zone",
              "default": "Joshimath Subsidence Zone",
              "title": "Location Name"
            },
            "description": "Target zone"
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 30.5574,
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 79.5662,
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Insar Subsidence Profile Api Futuristic Insar Subsidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/futuristic/debate": {
      "get": {
        "tags": [
          "Futuristic Earth Intelligence"
        ],
        "summary": "Get Agent Debate Protocol",
        "description": "Executes or retrieves an autonomous Graph-of-Thought multi-agent debate session.\nDemonstrates Optical vs SAR Radar cross-sensor arbitration without human intervention.",
        "operationId": "get_agent_debate_protocol_api_futuristic_debate_get",
        "parameters": [
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Scenario type: monsoon_flood, urban_shadow, or landslide",
              "default": "monsoon_flood",
              "title": "Scenario"
            },
            "description": "Scenario type: monsoon_flood, urban_shadow, or landslide"
          },
          {
            "name": "target_location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Brahmaputra Valley, Assam",
              "title": "Target Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDebateSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentDebateSession": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "target_location": {
            "type": "string",
            "title": "Target Location"
          },
          "dispute_topic": {
            "type": "string",
            "title": "Dispute Topic"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/DebateTurn"
            },
            "type": "array",
            "title": "Turns"
          },
          "consensus_reached": {
            "type": "boolean",
            "title": "Consensus Reached"
          },
          "final_verdict": {
            "type": "string",
            "title": "Final Verdict"
          },
          "final_confidence": {
            "type": "number",
            "title": "Final Confidence"
          },
          "overruled_sensor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overruled Sensor"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "scenario",
          "target_location",
          "dispute_topic",
          "turns",
          "consensus_reached",
          "final_verdict",
          "final_confidence"
        ],
        "title": "AgentDebateSession"
      },
...
      "DebateTurn": {
        "properties": {
          "turn_id": {
            "type": "integer",
            "title": "Turn Id"
          },
          "speaker": {
            "type": "string",
            "title": "Speaker"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "avatar_color": {
            "type": "string",
            "title": "Avatar Color"
          },
          "argument": {
            "type": "string",
            "title": "Argument"
          },
          "sensor_metric": {
            "type": "string",
            "title": "Sensor Metric"
          },
          "confidence_shift": {
            "type": "number",
            "title": "Confidence Shift"
          },
          "timestamp_offset_ms": {
            "type": "integer",
            "title": "Timestamp Offset Ms"
          }
        },
        "type": "object",
        "required": [
          "turn_id",
          "speaker",
          "role",
          "avatar_color",
          "argument",
          "sensor_metric",
          "confidence_shift",
          "timestamp_offset_ms"
        ],
        "title": "DebateTurn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}